app/phy/cpmGasprices/GoogleAppEngine/dev/requirements.txt
Flask==2.0.2
matplotlib-inline==0.1.3
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
requests==2.26.0
google-api-core==2.3.1
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
pandas==1.3.5
matplotlib==3.5.1
fsspec==2021.11.1
gcsfs==2021.11.1
gunicorn==20.1.0
